+++ physics/physics.vcxproj
@@ -161,7 +161,8 @@
     <ClInclude Include="discrete_trajectory.hpp" />
     <ClInclude Include="discrete_trajectory_body.hpp" />
     <ClInclude Include="dynamic_frame.hpp" />
-    <ClInclude Include="dynamic_frame_body.hpp" />
+    <ClInclude Include="rigid_motion.hpp" />
+    <ClInclude Include="rigid_motion_body.hpp" />
     <ClInclude Include="ephemeris.hpp" />
     <ClInclude Include="ephemeris_body.hpp" />
     <ClInclude Include="forkable.hpp" />
@@ -187,7 +188,7 @@
     <ClCompile Include="continuous_trajectory_test.cpp" />
     <ClCompile Include="degrees_of_freedom_test.cpp" />
     <ClCompile Include="discrete_trajectory_test.cpp" />
-    <ClCompile Include="dynamic_frame_test.cpp" />
+    <ClCompile Include="rigid_motion_test.cpp" />
     <ClCompile Include="ephemeris_test.cpp" />
     <ClCompile Include="forkable_test.cpp" />
     <ClCompile Include="solar_system_test.cpp" />
