+++ physics/rigid_motion_body.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
-#include "physics/dynamic_frame.hpp"
+#include "physics/rigid_motion.hpp"
 
 namespace principia {
 namespace physics {
-template <typename FromFrame, typename ToFrame>
+
+template<typename FromFrame, typename ToFrame>
 RigidMotion<FromFrame, ToFrame>::RigidMotion(
     RigidTransformation<FromFrame, ToFrame> const& rigid_transformation,
     AngularVelocity<ToFrame> const& rotation,
@@ -13,39 +14,37 @@ RigidMotion<FromFrame, ToFrame>::RigidMotion(
       rotation_(rotation),
       translation_(translation) {}
 
-template <typename FromFrame, typename ToFrame>
+template<typename FromFrame, typename ToFrame>
 RigidTransformation<FromFrame, ToFrame> const&
 RigidMotion<FromFrame, ToFrame>::rigid_transformation() const {
   return rigid_transformation_;
 }
 
-template <typename FromFrame, typename ToFrame>
+template<typename FromFrame, typename ToFrame>
 OrthogonalMap<FromFrame, ToFrame> const&
 RigidMotion<FromFrame, ToFrame>::orthogonal_map() const {
   return rigid_transformation_.linear_map();
 }
 
-template <typename FromFrame, typename ToFrame>
+template<typename FromFrame, typename ToFrame>
 DegreesOfFreedom<ToFrame> RigidMotion<FromFrame, ToFrame>::operator()(
     DegreesOfFreedom<FromFrame> const& degrees_of_freedom) const {
-  Position<ToFrame> const result_position =
-      rigid_transformation_(degrees_of_freedom.position());
-  return {result_position,
+  return {rigid_transformation_(degrees_of_freedom.position()),
           orthogonal_map()(degrees_of_freedom.velocity()) + translation_ +
               rotation_ * orthogonal_map()(degrees_of_freedom.position() -
                                            FromFrame::origin) / Radian};
 }
 
-template <typename FromFrame, typename ToFrame>
-RigidMotion<ToFrame, FromFrame> RigidMotion<FromFrame, ToFrame>::Inverse()
-    const {
+template<typename FromFrame, typename ToFrame>
+RigidMotion<ToFrame, FromFrame>
+RigidMotion<FromFrame, ToFrame>::Inverse() const {
   return RigidMotion<ToFrame, FromFrame>(
       rigid_transformation_.Inverse(),
       -orthogonal_map().Inverse()(rotation_),
       -orthogonal_map().Inverse()(translation_));
 }
 
-template <typename FromFrame, typename ThroughFrame, typename ToFrame>
+template<typename FromFrame, typename ThroughFrame, typename ToFrame>
 RigidMotion<FromFrame, ToFrame> operator*(
     RigidMotion<ThroughFrame, ToFrame> const& left,
     RigidMotion<FromFrame, ThroughFrame> const& right) {
