+++ physics/rigid_motion_test.cpp
@@ -1,11 +1,11 @@
-﻿#include "physics/dynamic_frame.hpp"
+﻿#include "physics/rigid_motion.hpp"
 
 #include "geometry/frame.hpp"
 #include "geometry/permutation.hpp"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
-#include "quantities/numbers.hpp"
 #include "quantities/named_quantities.hpp"
+#include "quantities/numbers.hpp"
 #include "quantities/si.hpp"
 #include "testing_utilities/almost_equals.hpp"
 #include "testing_utilities/componentwise.hpp"
